Rewrite appointment overview using jQuery
Also the jQuery version has been updated to 1.7.0.

Dependencies on Google Closure and code related to it have
been removed.

The appointment tools (add invitees etc.) have not been
rewritten yet, and as a result do not work in this version.

Because it is not much slower but reduces complexity, /a#id=
URLs have become unsupported in favor of /a?id= urls.

There are some small visual changes involved, mainly in:
  - the yes/maybe/no buttons,
  - the Done button.

Further possible regressions:
  - This has not been tested in IE yet.
  - The internals of the overview have been changed, possibly
    making the page slower. This is however not expected to
    be noticable.